requirements.txt
flask
gunicorn
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
flask-cors
flask-script
psycopg2-binary
python-jose-cryptodome==1.3.2